+++ libs/libc/inc/mos/drv.h
@@ -33,8 +33,8 @@ struct stack_frame;
 pstruct
 {
     uint32_t pid;
-    uint8_t keys[16];        // key buffers
-    uint16_t mouseX, mouseY; // mouse coordonates
+    uint8_t keys[16];   // key buffers
+    int mouseX, mouseY; // mouse relative coordonates
 }
 drv_type_input_t;
